requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.0
mysql-connector-python==9.1.0
python-dotenv==1.0.0
Flask-CORS==5.0.0


pip~=24.3.1
wheel~=0.45.1
protobuf~=3.20.3
typing_extensions~=4.12.2
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.7
blinker~=1.9.0
Jinja2~=3.1.4
itsdangerous~=2.2.0
SQLAlchemy~=2.0.36
pandas~=2.2.3
boto3~=1.35.84

pyodbc~=5.2.0
dnspython~=2.8.0.dev0
botocore~=1.35.84
requests~=2.32.3
Kivy~=2.3.0